daemon: cleaned up compiler warnings
raised warning level to /Wall
changed nfs41_file_info.owner, owner_group to char[] to avoid casting

+++ daemon/readdir.c
@@ -140,7 +140,7 @@ static void readdir_copy_dir_info(
     IN nfs41_readdir_entry *entry,
     IN PFILE_DIR_INFO_UNION info)
 {
-    info->fdi.FileIndex = entry->attr_info.fileid;
+    info->fdi.FileIndex = (ULONG)entry->attr_info.fileid;
     nfs_time_to_file_time(&entry->attr_info.time_create,
         &info->fdi.CreationTime);
     nfs_time_to_file_time(&entry->attr_info.time_access,
